# feat: simplify and speedup feature_governance.py test by generating less blocks