## Issue being fixed or feature implemented
Too many blocks are generated in a functional tests `feature_governance.py`

## What was done?
1. Activate v20 faster for `feature_governance.py`, drop generation of useless blocks, simplify code.


2. Added extra logging for this test by converting comments to logs.


## How Has This Been Tested?
Run `test/functional/feature_governance.py`
Locally improvement of performance is less than 5 seconds.
`tsan` jobs on CI:
```
204/264 - feature_governance.py --legacy-wallet passed, Duration: 243 s
208/264 - feature_governance.py --descriptors passed, Duration: 270 s
```
↑ [old version](https://gitlab.com/dashpay/dash/-/jobs/7805625816) vs [new version](https://gitlab.com/dashpay/dash/-/jobs/7819897536) ↓:
```
193/264 - feature_governance.py --legacy-wallet passed, Duration: 220 s
199/264 - feature_governance.py --descriptors passed, Duration: 240 s
```

Most of the time is spent in `feature_governance.py` on stage "let all fulfilled requests expire for re-sync to work correctly" so that's a very minor optimization.

## Breaking Changes
N/A

## Checklist:
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone